release: Allow ARCH to be overridden
It is possible to pre-set/export ARCH which will be detected by
the release script.  This should allow versions destined for
different architectures to be built and pushed from one machine.

Note: AArch64 is still in bootstrap mode, thus it is currently
unable to build the x86_64 images since those use 'crux:latest'
as a build platform.  This will be fixed once AArch64
'crux:latest' has been released.

Signed-off-by: Lee Jones <lee.jones@linaro.org>